Added CateogryExists and SubcategoryExists methods

+++ OnlineLibrary/OnlineLibrary.Services/Concrete/CategoryService.cs
@@ -163,12 +163,16 @@ namespace OnlineLibrary.Services.Concrete
 
         public void DeleteBookCategory(int cateogryId)
         {
+            if (!CategoryExists(cateogryId))
+            {
+                throw new BookCategoryNotFoundException("Book Category does't exist");
+            }
             if (!IsCategoryRemovable(cateogryId))
             {
                 throw new BookCategoryIsNotRemovableException("Book category has subcategories");
             }
 
-            var removedCategory = _dbContext.Categories.FirstOrDefault(c => c.Id == cateogryId);
+            var removedCategory = _dbContext.Categories.Find(cateogryId);
             _dbContext.Categories.Remove(removedCategory);
             _dbContext.SaveChanges();
 
@@ -180,14 +184,27 @@ namespace OnlineLibrary.Services.Concrete
             return !_dbContext.SubCategories.Any(sc => sc.CategoryId == categoryId);
         }
 
+        public bool CategoryExists(int categoryId)
+        {
+            if (_dbContext.Categories.Find(categoryId) != null)
+            {
+                return true;
+            }
+            return false;
+        }
+
         public void DeleteBookSubcategory(int subcategoryId)
         {
-            if(!IsSubcategoryRemovable(subcategoryId))
+            if (!SubcategoryExists(subcategoryId))
+            {
+                throw new BookSubcategoryNotFoundException("Subcategory doesn't exist");
+            }
+            if (!IsSubcategoryRemovable(subcategoryId))
             {
                 throw new BookSubcateogryIsNotRemovableException("Subcategory has books");
             }
 
-            var removedSubcategory = _dbContext.SubCategories.FirstOrDefault(sc => sc.Id == subcategoryId);
+            var removedSubcategory = _dbContext.SubCategories.Find(subcategoryId);
             _dbContext.SubCategories.Remove(removedSubcategory);
             _dbContext.SaveChanges();
         }
@@ -196,6 +213,14 @@ namespace OnlineLibrary.Services.Concrete
         {
             return !_dbContext.Books.Any(b => b.SubCategories.Any(sc => sc.Id == subcategoryId));
         }
-        
+
+        public bool SubcategoryExists(int subcategoryId)
+        {
+            if (_dbContext.SubCategories.Find(subcategoryId) != null)
+            {
+                return true;
+            }
+            return false;
+        }        
     }
 }
